backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
boto3>=1.34.129
requests-oauthlib>=2.0.0
cryptography>=42.0.8
python-dotenv>=1.0.1
pymongo==4.5.0
pydantic>=2.6.4
email-validator>=2.2.0
pyjwt>=2.10.1
passlib>=1.7.4
tzdata>=2024.2
motor==3.3.1
pytest>=8.0.0
black>=24.1.1
isort>=5.13.2
flake8>=7.0.0
mypy>=1.8.0
python-jose>=3.3.0
requests>=2.31.0
pandas>=2.2.0
numpy>=1.26.0
python-multipart>=0.0.9
jq>=1.6.0
typer>=0.9.0
requests-cache>=1.2.0
scipy>=1.11.0
ta>=0.11.0

# Backtesting and Historical Data
ccxt>=4.4.96
matplotlib>=3.10.3

# Additional Security Dependencies
bcrypt==4.1.2
argon2-cffi==23.1.0
pyotp==2.9.0
qrcode==7.4.2
pillow==10.1.0

# Monitoring & Logging
structlog==23.2.0
sentry-sdk==1.39.2

# Google AI for chat functionality
google-generativeai==0.3.0
openai==1.25.0